backend/requirements.txt
Flask
flask-cors
gunicorn
python-dotenv
paho-mqtt